Smartify monsters and add win/loss conditions
Currently the monsters will chase the player from all over the map to try to get
to them. They also aren't smart enough so that they surround the player. Uses a
Dijkstra algorithm implementation for the pathfinding. The win/loss states pop
up a new screen depending on which one it is with some flavor text.